Forgot.
Completely forgot that the broken paltforms were not working after i submitted... quick fix by splicing the platform if it rolls the 30% chance and slo making the doodler jump. Also forgot to reset the score multiplier at the result screen, and also the doodler y position att 350 so it does not spawn att y=600.